% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu.R
\name{fu}
\alias{fu}
\title{Might correct you, when you fucked up.}
\usage{
fu(error = geterrmessage(), confirm = FALSE)
}
\arguments{
\item{error}{[\code{character(1)}]\cr
\code{fuck()} will automatically use the last error message but you can also provide your own.}

\item{confirm}{[\code{logical(1)}]\cr
Should this problem be solved without further confirmation?}
}
\description{
Might correct you, when you fucked up.
}
